Refactoring and support new features
 - refactor renderer
 - refecror document builder
 - add support for blockquotes
 - add support for code blocks
 - add support for cell margin
 - use patched version of gofpdf with text block background and fixed
position of background
