+++ go.mod
@@ -3,6 +3,6 @@ module github.com/raykov/mdtopdf
 go 1.16
 
 require (
-	github.com/jung-kurt/gofpdf v1.16.2
-	github.com/yuin/goldmark v1.4.0
+	github.com/raykov/gofpdf v1.16.7
+	github.com/yuin/goldmark v1.4.1
 )
